`fn save_tmvs_c`: Remove redundant `as` casts.

+++ src/refmvs.rs
@@ -1477,9 +1477,9 @@ unsafe extern "C" fn save_tmvs_c(
         let mut x = col_start8;
         while x < col_end8 {
             let cand_b = (*b.offset((x * 2 + 1) as isize)).0;
-            let bw8 = dav1d_block_dimensions[cand_b.bs as usize][0] as c_int + 1 >> 1;
-            if cand_b.r#ref.r#ref[1] as c_int > 0
-                && ref_sign[(cand_b.r#ref.r#ref[1] as c_int - 1) as usize] as c_int != 0
+            let bw8 = dav1d_block_dimensions[cand_b.bs as usize][0] + 1 >> 1;
+            if cand_b.r#ref.r#ref[1] > 0
+                && ref_sign[cand_b.r#ref.r#ref[1] as usize - 1] != 0
                 && cand_b.mv.mv[1].y.abs() | cand_b.mv.mv[1].x.abs() < 4096
             {
                 let mut n = 0;
@@ -1494,8 +1494,8 @@ unsafe extern "C" fn save_tmvs_c(
                     n += 1;
                     x += 1;
                 }
-            } else if cand_b.r#ref.r#ref[0] as c_int > 0
-                && ref_sign[(cand_b.r#ref.r#ref[0] as c_int - 1) as usize] as c_int != 0
+            } else if cand_b.r#ref.r#ref[0] > 0
+                && ref_sign[cand_b.r#ref.r#ref[0] as usize - 1] != 0
                 && cand_b.mv.mv[0].y.abs() | cand_b.mv.mv[0].x.abs() < 4096
             {
                 let mut n = 0;
